wificfg: correct some signed size declarations.

+++ extras/wificfg/wificfg.c
@@ -66,7 +66,7 @@ static int read_crlf_line(int s, char *buf, size_t len)
 
     do {
         char c;
-        int r = read(s, &c, 1);
+        ssize_t r = read(s, &c, 1);
 
         /* Expecting a known terminator so fail on EOF. */
         if (r <= 0)
@@ -91,7 +91,7 @@ static int read_crlf_line(int s, char *buf, size_t len)
     return num;
 }
 
-int wificfg_form_name_value(int s, bool *valp, size_t *rem, char *buf, size_t len)
+ssize_t wificfg_form_name_value(int s, bool *valp, size_t *rem, char *buf, size_t len)
 {
     size_t num = 0;
 
@@ -100,7 +100,7 @@ int wificfg_form_name_value(int s, bool *valp, size_t *rem, char *buf, size_t le
             break;
 
         char c;
-        int r = read(s, &c, 1);
+        ssize_t r = read(s, &c, 1);
 
         /* Expecting a known number of characters so fail on EOF. */
         if (r <= 0) return -1;
@@ -233,7 +233,7 @@ static const struct {
 
 static wificfg_method intern_http_method(char *str)
 {
-    int i;
+    size_t i;
     for (i = 0;  i < sizeof(method_table) / sizeof(method_table[0]); i++) {
         if (!strcmp(str, method_table[i].str))
             return method_table[i].method;
@@ -264,7 +264,7 @@ static const struct {
 
 static http_header intern_http_header(char *str)
 {
-    int i;
+    size_t i;
     for (i = 0;  i < sizeof(http_header_table) / sizeof(http_header_table[0]); i++) {
         if (!strcmp(str, http_header_table[i].str))
             return http_header_table[i].name;
@@ -282,7 +282,7 @@ static const struct {
 
 static wificfg_content_type intern_http_content_type(char *str)
 {
-    int i;
+    size_t i;
     for (i = 0;  i < sizeof(content_type_table) / sizeof(content_type_table[0]); i++) {
         if (!strcmp(str, content_type_table[i].str))
             return content_type_table[i].type;
@@ -308,13 +308,13 @@ static char *skip_to_whitespace(char *string)
     return string;
 }
 
-int wificfg_write_string(int s, const char *str)
+ssize_t wificfg_write_string(int s, const char *str)
 {
-    int res = write(s, str, strlen(str));
+    ssize_t res = write(s, str, strlen(str));
     return res;
 }
 
-int wificfg_write_string_chunk(int s, const char *str, char *buf, size_t len)
+ssize_t wificfg_write_string_chunk(int s, const char *str, char *buf, size_t len)
 {
     size_t str_len = strlen(str);
 
@@ -356,7 +356,7 @@ int wificfg_write_string_chunk(int s, const char *str, char *buf, size_t len)
     /* Else too big for the buffer. */
     char size_buf[8];
     size_t size_len = snprintf(size_buf, sizeof(size_buf), "%x\r\n", str_len);
-    int res = write(s, size_buf, size_len);
+    ssize_t res = write(s, size_buf, size_len);
     if (res != size_len) {
         return res;
     }
@@ -367,7 +367,7 @@ int wificfg_write_string_chunk(int s, const char *str, char *buf, size_t len)
     return write(s, size_buf + size_len - 2, 2);
 }
 
-int wificfg_write_chunk_end(int s)
+ssize_t wificfg_write_chunk_end(int s)
 {
     return wificfg_write_string(s, "0\r\n\r\n");
 }
@@ -436,12 +436,12 @@ static const struct {
 
 static form_name intern_form_name(char *str)
 {
-     int i;
-     for (i = 0;  i < sizeof(form_name_table) / sizeof(form_name_table[0]); i++) {
-         if (!strcmp(str, form_name_table[i].str))
-             return form_name_table[i].name;
-     }
-     return FORM_NAME_NONE;
+    size_t i;
+    for (i = 0;  i < sizeof(form_name_table) / sizeof(form_name_table[0]); i++) {
+        if (!strcmp(str, form_name_table[i].str))
+            return form_name_table[i].name;
+    }
+    return FORM_NAME_NONE;
 }
 
 
@@ -759,7 +759,7 @@ static int handle_wificfg_index_post(int s, wificfg_method method,
     bool valp = false;
 
     while (rem > 0) {
-        int r = wificfg_form_name_value(s, &valp, &rem, buf, len);
+        ssize_t r = wificfg_form_name_value(s, &valp, &rem, buf, len);
 
         if (r < 0) {
             break;
@@ -770,7 +770,7 @@ static int handle_wificfg_index_post(int s, wificfg_method method,
         form_name name = intern_form_name(buf);
 
         if (valp) {
-            int r = wificfg_form_name_value(s, NULL, &rem, buf, len);
+            ssize_t r = wificfg_form_name_value(s, NULL, &rem, buf, len);
             if (r < 0) {
                 break;
             }
@@ -1139,7 +1139,7 @@ static int handle_wifi_ap_post(int s, wificfg_method method,
     uint8_t dns_enable = 0;
 
     while (rem > 0) {
-        int r = wificfg_form_name_value(s, &valp, &rem, buf, len);
+        ssize_t r = wificfg_form_name_value(s, &valp, &rem, buf, len);
 
         if (r < 0) {
             break;
@@ -1150,7 +1150,7 @@ static int handle_wifi_ap_post(int s, wificfg_method method,
         form_name name = intern_form_name(buf);
 
         if (valp) {
-            int r = wificfg_form_name_value(s, NULL, &rem, buf, len);
+            ssize_t r = wificfg_form_name_value(s, NULL, &rem, buf, len);
             if (r < 0) {
                 break;
             }
@@ -1679,7 +1679,7 @@ static void server_task(void *pvParameters)
                 size_t len;
                 for (len = 0; len < 4096; len++) {
                     char c;
-                    int res = read(s, &c, 1);
+                    ssize_t res = read(s, &c, 1);
                     if (res != 1) break;
                 }
 
@@ -1730,7 +1730,7 @@ static void dns_task(void *pvParameters)
         char buffer[96];
         struct sockaddr src_addr;
         socklen_t src_addr_len = sizeof(src_addr);
-        size_t count = recvfrom(fd, buffer, sizeof(buffer), 0, (struct sockaddr*)&src_addr, &src_addr_len);
+        ssize_t count = recvfrom(fd, buffer, sizeof(buffer), 0, (struct sockaddr*)&src_addr, &src_addr_len);
 
         /* Drop messages that are too large to send a response in the buffer */
         if (count > 0 && count <= sizeof(buffer) - 16 && src_addr.sa_family == AF_INET) {
